adicionando as controllers Chamado,Usuario, Maquininha

+++ src/main/java/br/com/meli/helpdesksimapi/service/MaquininhaService.java
@@ -1,12 +1,12 @@
 package br.com.meli.helpdesksimapi.service;
 
-import br.com.meli.helpdesksimapi.model.Balcao;
 import br.com.meli.helpdesksimapi.model.Maquininha;
 import br.com.meli.helpdesksimapi.repository.MaquininhaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class MaquininhaService {
@@ -22,7 +22,7 @@ public class MaquininhaService {
         return maquininhaRepository.findAll();
     }
 
-    private Maquininha buscarMaquininhaPorId(Long id) {
+    public Maquininha buscarMaquininhaPorId(Long id) {
         return maquininhaRepository.findById(id).orElse(null);
     }
 
@@ -30,7 +30,13 @@ public class MaquininhaService {
         return maquininhaRepository.save(maquininha);
     }
 
-    public void excluirMaquininha(Long id) {
-        maquininhaRepository.deleteById(id);
+    public boolean deletarMaquininha(Long id) {
+        Optional<Maquininha> maquininha = maquininhaRepository.findById(id);
+        if( maquininha.isPresent() ) {
+            maquininhaRepository.deleteById(id);
+            return true;
+        }else {
+            return false;
+        }
     }
 }
